weight is now a byte instead of a float

+++ PointCloudEngine/DataStructures.h
@@ -114,11 +114,11 @@ namespace PointCloudEngine
         Vector3 position;
 
         // The different cluster mean normals and colors in object space calculated by k-means algorithm with k=6
-        // Each weights is the percentage of points assigned to this cluster
-        // TODO: Optimize weights byte size
+        // Each weights is the percentage of points assigned to this cluster (0=0%, 255=100%)
+        // TODO: One weight can be omitted because the sum of all weights is always 1
         PolarNormal normals[6];
         Color16 colors[6];
-        float weights[6];
+        byte weights[6];
 
         // Width of the whole cube
         float size;
